Improved image type support, fixed z-spacing
When Bio-Formats returns a BufferedImage with TYPE_CUSTOM, the transfer type for the raster is now checked to see if it is something that can be readily handled (here, TYPE_BYTE_GRAY or TYPE_USHORT_GRAY).

Also, the zSpacing value is set even if less than 1 (previously it was Double.NaN in this case).

Version number increased and .gitignore also updated to allow gradle.properties to be set locally.